improve the performance of savePosts method
修改 改善保存通知方法的性能，新方法名为savePosts
修改 doUpdatePosts时，如果无新通知直接返回0，避免白开数据库

+++ src/org/orange/querysystem/PostsActivity.java
@@ -62,7 +62,7 @@ public class PostsActivity extends FragmentActivity{
 		mWebUpdaterToDB = new PostUpdater(this);
 		mWebUpdaterToDB.setOnPostExecuteListener(new OnPostUpdateListener() {
 			@Override
-			public void onPostUpdate(int numberOfInsertedPosts, boolean mandatorily) {
+			public void onPostUpdate(long numberOfInsertedPosts, boolean mandatorily) {
 				if(mandatorily || numberOfInsertedPosts>0){
 					if(numberOfInsertedPosts > 0){
 						loadPosts();
